feat(roll): add model backup and review functionality
- Introduced model_backup.py for compressing and decompressing model run directories, enhancing data management.
- Added model_review.py to facilitate detailed analysis of model predictions, improving the review process.
- Refactored ModelCLI to delegate review logic to ModelReviewHelper, reducing complexity and improving maintainability.
